# api/openapi.yaml
openapi: 3.0.3
info:
  title: Task & Project Management System API
  description: >
    Documents the REST API exactly as implemented in `backend/` today — not an
    idealized target contract. This is the hardened, post-review version of
    the backend: every controller now validates its input, returns clean
    JSON errors, uses response DTOs (never raw JPA entities), sets correct
    HTTP status codes, and enforces role-based access control. Remaining
    gaps/quirks are called out inline via `description` fields; see
    `api/README.md` for the summary list.
  version: "0.2.0"
servers:
  - url: http://localhost:8080
    description: Local (via docker-compose or `mvn spring-boot:run`)
security:
  - bearerAuth: []
...
paths:
...
  /api/milestones:
    get:
      tags: [Milestones]
      summary: List all milestones
      description: Any authenticated user — not role-gated.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MilestoneResponse"
    post:
      tags: [Milestones]
      summary: Create a milestone
      description: "`ADMINISTRATOR` / `PROJECT_MANAGER` / `TEAM_LEADER`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: "`projectId` does not match any project."
          $ref: "#/components/responses/NotFound"
  /api/milestones/{id}:
    get:
      tags: [Milestones]
      summary: Get a milestone by ID
      description: Any authenticated user — not role-gated.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Milestones]
      summary: Replace a milestone
      description: >
        `ADMINISTRATOR` / `PROJECT_MANAGER` / `TEAM_LEADER`. Full-replace
        semantics — no `PATCH`, and no dedicated "update progress/status"
        endpoint. No server-side recompute of the parent project's
        `progress` when a milestone changes.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MilestoneRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MilestoneResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: Milestone not found, or `projectId` does not match any project.
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Milestones]
      summary: Delete a milestone
      description: "`ADMINISTRATOR` / `PROJECT_MANAGER` / `TEAM_LEADER`."
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted — no content.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/milestones/project/{projectId}:
    get:
      tags: [Milestones]
      summary: List milestones for a project
      description: >
        Any authenticated user — not role-gated. An unknown `projectId`
        returns an empty list rather than a `404`.
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MilestoneResponse"
  /api/tasks:
    get:
      tags: [Tasks]
      summary: List all tasks
      description: Any authenticated user — not role-gated.
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
    post:
      tags: [Tasks]
      summary: Create a task
      description: "`ADMINISTRATOR` / `PROJECT_MANAGER` / `TEAM_LEADER`."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          description: >
            Bean Validation failure, or a DB trigger rejection (e.g. the
            task's `dueDate` is later than its project's `endDate`) —
            both come back as this same `ErrorResponse` shape, cleaned of
            Postgres's `ERROR:`/`Where:` noise by `GlobalExceptionHandler`.
          $ref: "#/components/responses/BadRequest"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: >
            `projectId`, `milestoneId`, or `createdById` does not match any
            row.
          $ref: "#/components/responses/NotFound"
  /api/tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get a task by ID
      description: Any authenticated user — not role-gated.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Tasks]
      summary: Replace a task
      description: >
        No `@PreAuthorize` — open to any authenticated role, but
        **ownership-scoped rather than role-gated** (`TaskService.updateTask`):
        `ADMINISTRATOR`/`PROJECT_MANAGER`/`TEAM_LEADER` may edit every field
        of any task, full-replace as usual (no `PATCH`). A `TEAM_MEMBER` may
        only call this for a task they are currently a `task_assignees` row
        for — otherwise `403`. Even when allowed, only `status`/`progress`
        from the request body take effect for a `TEAM_MEMBER`; every other
        field (title, project, milestone, dates, etc.) is silently left
        unchanged, even though the client is still expected to send the full
        `TaskRequest` body. A real status change (by anyone) fires a
        `TASK_STATUS_CHANGED` notification to every current assignee.
        `completedAt` is never auto-stamped — the caller must set it
        explicitly when marking a task `COMPLETED`.
      parameters:
        - $ref: "#/components/parameters/IdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaskResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "403":
          description: >
            Caller is a `TEAM_MEMBER` and is not currently assigned to this
            task.
          $ref: "#/components/responses/Forbidden"
        "404":
          description: >
            Task not found, or (for a full edit) `projectId`/`milestoneId`/
            `createdById` does not match any row.
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Tasks]
      summary: Delete a task
      description: "`ADMINISTRATOR` / `PROJECT_MANAGER` / `TEAM_LEADER`."
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "204":
          description: Deleted — no content.
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/tasks/project/{projectId}:
    get:
      tags: [Tasks]
      summary: List tasks in a project
      description: >
        Any authenticated user — not role-gated. An unknown `projectId`
        returns an empty list rather than a `404`.
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
  /api/tasks/milestone/{milestoneId}:
    get:
      tags: [Tasks]
      summary: List tasks under a milestone
      description: >
        Any authenticated user — not role-gated. An unknown `milestoneId`
        returns an empty list rather than a `404`.
      parameters:
        - name: milestoneId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
  /api/tasks/status/{status}:
    get:
      tags: [Tasks]
      summary: List tasks by status
      description: >
        Any authenticated user — not role-gated. `status` is a free-text path
        segment (`TaskRepository.findByStatus`), not validated against the
        `TaskStatus` enum below — an unrecognized value just returns an empty
        list rather than a `400`.
      parameters:
        - name: status
          in: path
          required: true
          schema:
            type: string
          example: IN_PROGRESS
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskResponse"
...
  /api/task-assignees/task/{taskId}:
    get:
      tags: [Task Assignees]
      summary: List assignees of a task
      description: >
        Any authenticated user — not role-gated. An unknown `taskId` returns
        an empty list rather than a `404`.
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskAssigneeResponse"
...
  /api/task-dependencies/task/{taskId}:
    get:
      tags: [Task Dependencies]
      summary: List what a task depends on
      description: >
        Any authenticated user — not role-gated. An unknown `taskId` returns
        an empty list rather than a `404`.
      parameters:
        - name: taskId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaskDependencyResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Required on every endpoint except `GET /api/health` and
        `POST /api/auth/login`. Authorization beyond "is logged in" is
        role-based via `@PreAuthorize` (`config/SecurityConfig.java`,
        `@EnableMethodSecurity`) — the JWT's custom `role` claim is mapped to
        a `ROLE_<name>` granted authority by a custom
        `JwtAuthenticationConverter`. Policy is coarse role gates per
        resource (see each operation's description and
        `backend/README.md#security`), with one ownership-scoped exception:
        `PUT /api/tasks/{id}` for a `TEAM_MEMBER` caller. Every `GET`
        endpoint, `PUT /api/users/me`, and every `/api/notifications`
        endpoint have no role gate at all beyond a valid JWT.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64
  responses:
    NotFound:
      description: >
        No row matches the given id (or a referenced foreign id in the
        request body). `GlobalExceptionHandler` maps `NotFoundException` to a
        clean `404` JSON body — never a raw stack trace.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2026-09-12T10:15:30Z"
            status: 404
            error: "Not Found"
            message: "Task not found"
            fieldErrors: null
...
    BadRequest:
      description: >
        Either a Bean Validation failure (`fieldErrors` populated — see
        `ValidationError`) or a DB constraint/trigger violation
        (`DataIntegrityViolationException`, `fieldErrors` is `null` and
        `message` is the cleaned-up constraint/trigger message, with
        Postgres's `ERROR:`/`Where:` noise stripped by
        `GlobalExceptionHandler.cleanPostgresMessage`). Both use the same
        `ErrorResponse` shape and status code, so a client must inspect
        `fieldErrors`/`error` to tell them apart.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          examples:
            validation:
              value:
                timestamp: "2026-09-12T10:15:30Z"
                status: 400
                error: "Validation Failed"
                message: "One or more fields are invalid"
                fieldErrors:
                  title: "must not be blank"
            dataIntegrity:
              value:
                timestamp: "2026-09-12T10:15:30Z"
                status: 400
                error: "Data Integrity Violation"
                message: "Task due_date (2026-09-18) cannot be later than its project end_date (2026-07-01)"
                fieldErrors: null
...
    Forbidden:
      description: >
        Authenticated, but the caller's role (or, for
        `PUT /api/tasks/{id}`, assignment status) doesn't permit this
        operation. Mapped from Spring Security's `AccessDeniedException`.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            timestamp: "2026-09-12T10:15:30Z"
            status: 403
            error: "Forbidden"
            message: "You do not have permission to perform this action"
...
  schemas:
...
    ErrorResponse:
      type: object
      description: >
        The one error shape used across the whole API
        (`exception/ErrorResponse.java`, a record) — every handled failure
        path in `GlobalExceptionHandler` (404/400/401/403/429/500) returns
        this, never a raw stack trace or Spring's undocumented default body.
      required: [timestamp, status, error, message]
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
          example: 404
        error:
          type: string
          example: "Not Found"
        message:
          type: string
          example: "Task not found"
        fieldErrors:
          type: object
          nullable: true
          additionalProperties:
            type: string
          description: >
            Populated only for Bean Validation failures (one entry per
            invalid field); `null` for every other error type.

    # ---------- shared enum-ish string fields ----------
    # Not a real Java enum type (@Enumerated) anywhere — every one of these is
    # a plain String column, validated on write via @Pattern (Bean
    # Validation) against the same set of values the Postgres CHECK
    # constraint in database/init/01-init.sql allows. A request with an
    # invalid value now fails cleanly as a 400 (ValidationError) rather than
    # reaching the database at all.
    AccountStatus:
      type: string
      enum: [ACTIVE, INACTIVE, SUSPENDED]
      description: "`users.account_status`. Defaults to `ACTIVE` if omitted on create."
    ProjectPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, CRITICAL]
      description: "`projects.priority`. Defaults to `MEDIUM` if omitted on create."
    ProjectStatus:
      type: string
      enum: [PLANNING, IN_PROGRESS, ON_HOLD, COMPLETED, CANCELLED]
      description: "`projects.status`. Defaults to `PLANNING` if omitted on create."
...
    MilestoneStatus:
      type: string
      enum: [PENDING, IN_PROGRESS, COMPLETED]
      description: "`milestones.status`. Defaults to `PENDING` if omitted on create."
    TaskPriority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
      description: "`tasks.priority`. Defaults to `MEDIUM` if omitted on create."
    TaskStatus:
      type: string
      enum: [TO_DO, IN_PROGRESS, IN_REVIEW, COMPLETED, CANCELLED]
      description: "`tasks.status`. Defaults to `TO_DO` if omitted on create."
...
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fullName:
          type: string
        username:
          type: string
        email:
          type: string
        gender:
          type: string
          nullable: true
        dateOfBirth:
          type: string
          format: date
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        profilePhotoUrl:
          type: string
          nullable: true
        position:
          type: string
          nullable: true
        department:
          type: string
          nullable: true
        role:
          type: string
          description: Role name, e.g. `TEAM_MEMBER`.
        roleDescription:
          type: string
          nullable: true
        accountStatus:
          $ref: "#/components/schemas/AccountStatus"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ProjectResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        projectCode:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        manager:
          $ref: "#/components/schemas/UserResponse"
          description: Nested `UserResponse` DTO — never the raw entity, no `passwordHash`.
        priority:
          $ref: "#/components/schemas/ProjectPriority"
        status:
          $ref: "#/components/schemas/ProjectStatus"
        progress:
          type: number
          description: 0–100, DB-enforced (`NUMERIC(5,2) CHECK BETWEEN 0 AND 100`) and Bean-Validation-enforced on write (`@DecimalMin`/`@DecimalMax`).
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    MilestoneResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        project:
          $ref: "#/components/schemas/ProjectResponse"
        title:
          type: string
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/MilestoneStatus"
        progress:
          type: number
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    MilestoneRequest:
      type: object
      required: [projectId, title, dueDate]
      properties:
        projectId:
          type: integer
          format: int64
          description: Must reference an existing project, or the request fails with `404`.
        title:
          type: string
          maxLength: 200
        description:
          type: string
          nullable: true
        dueDate:
          type: string
          format: date
        status:
          $ref: "#/components/schemas/MilestoneStatus"
        progress:
          type: number
          minimum: 0
          maximum: 100

    # ---------- tasks ----------
    TaskResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        project:
          $ref: "#/components/schemas/ProjectResponse"
        milestone:
          $ref: "#/components/schemas/MilestoneResponse"
          nullable: true
        title:
          type: string
        description:
          type: string
          nullable: true
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        startDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
        estimatedHours:
          type: number
          nullable: true
        progress:
          type: number
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Not auto-set — nothing stamps this when `status` becomes `COMPLETED`; the caller must set it explicitly.
        createdBy:
          $ref: "#/components/schemas/UserResponse"
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    TaskRequest:
      type: object
      required: [projectId, title, startDate, dueDate]
      properties:
        projectId:
          type: integer
          format: int64
          description: Must reference an existing project, or the request fails with `404`.
        milestoneId:
          type: integer
          format: int64
          nullable: true
          description: Optional — omit/null for a task with no milestone. Must reference an existing milestone if given.
        title:
          type: string
          maxLength: 200
        description:
          type: string
          nullable: true
        priority:
          $ref: "#/components/schemas/TaskPriority"
        status:
          $ref: "#/components/schemas/TaskStatus"
        startDate:
          type: string
          format: date
        dueDate:
          type: string
          format: date
          description: >
            A DB trigger rejects a `dueDate` later than the parent project's
            `endDate` — surfaces as a `400` (see `BadRequest`), not a raw
            `500`.
        estimatedHours:
          type: number
          nullable: true
          minimum: 0
        progress:
          type: number
          minimum: 0
          maximum: 100
        completedAt:
          type: string
          format: date-time
          nullable: true
          description: Not auto-set — caller must set this explicitly when marking a task done.
        createdById:
          type: integer
          format: int64
          nullable: true
          description: Optional — the user who created the task, if known. Must reference an existing user if given.

    # ---------- task assignees ----------
    TaskAssigneeResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        task:
          $ref: "#/components/schemas/TaskResponse"
        user:
          $ref: "#/components/schemas/UserResponse"
        assignedAt:
          type: string
          format: date-time
...
    TaskDependencyResponse:
      type: object
      description: No surrogate `id` — identity is the (task, dependsOnTask) pair.
      properties:
        task:
          $ref: "#/components/schemas/TaskResponse"
        dependsOnTask:
          $ref: "#/components/schemas/TaskResponse"
...